setup-tool/requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-openai
openai